+++ matrix-sdk-android/src/main/java/im/vector/matrix/android/internal/database/mapper/TimelineEventMapper.kt
@@ -17,29 +17,38 @@
 package im.vector.matrix.android.internal.database.mapper
 
 import im.vector.matrix.android.api.session.events.model.Event
+import im.vector.matrix.android.api.session.room.model.ReadReceipt
+
 import im.vector.matrix.android.api.session.room.timeline.TimelineEvent
 import im.vector.matrix.android.internal.database.model.TimelineEventEntity
+import javax.inject.Inject
 
-internal object TimelineEventMapper {
+internal class TimelineEventMapper @Inject constructor(private val readReceiptsSummaryMapper: ReadReceiptsSummaryMapper) {
 
-    fun map(timelineEventEntity: TimelineEventEntity): TimelineEvent {
-
+    fun map(timelineEventEntity: TimelineEventEntity, buildReadReceipts: Boolean = true, correctedReadReceipts: List<ReadReceipt>? = null): TimelineEvent {
+        val readReceipts = if (buildReadReceipts) {
+            correctedReadReceipts ?: timelineEventEntity.readReceipts
+                    ?.let {
+                        readReceiptsSummaryMapper.map(it)
+                    }
+        } else {
+            null
+        }
         return TimelineEvent(
                 root = timelineEventEntity.root?.asDomain()
-                        ?: Event("", timelineEventEntity.eventId),
+                       ?: Event("", timelineEventEntity.eventId),
                 annotations = timelineEventEntity.annotations?.asDomain(),
                 localId = timelineEventEntity.localId,
                 displayIndex = timelineEventEntity.root?.displayIndex ?: 0,
                 senderName = timelineEventEntity.senderName,
                 isUniqueDisplayName = timelineEventEntity.isUniqueDisplayName,
-                senderAvatar = timelineEventEntity.senderAvatar
+                senderAvatar = timelineEventEntity.senderAvatar,
+                readReceipts = readReceipts?.sortedByDescending {
+                    it.originServerTs
+                } ?: emptyList()
         )
     }
 
-}
-
-internal fun TimelineEventEntity.asDomain(): TimelineEvent {
-    return TimelineEventMapper.map(this)
 }