+++ matrix-sdk-android/src/main/java/im/vector/matrix/android/internal/session/room/read/DefaultReadService.kt
@@ -16,14 +16,20 @@
 
 package im.vector.matrix.android.internal.session.room.read
 
+import androidx.lifecycle.LiveData
+import androidx.lifecycle.Transformations
 import com.squareup.inject.assisted.Assisted
 import com.squareup.inject.assisted.AssistedInject
 import com.zhuinden.monarchy.Monarchy
 import im.vector.matrix.android.api.MatrixCallback
 import im.vector.matrix.android.api.auth.data.Credentials
+import im.vector.matrix.android.api.session.room.model.ReadReceipt
 import im.vector.matrix.android.api.session.room.read.ReadService
+import im.vector.matrix.android.internal.database.RealmLiveData
+import im.vector.matrix.android.internal.database.mapper.ReadReceiptsSummaryMapper
 import im.vector.matrix.android.internal.database.model.ChunkEntity
 import im.vector.matrix.android.internal.database.model.ReadReceiptEntity
+import im.vector.matrix.android.internal.database.model.ReadReceiptsSummaryEntity
 import im.vector.matrix.android.internal.database.query.find
 import im.vector.matrix.android.internal.database.query.findLastLiveChunkFromRoom
 import im.vector.matrix.android.internal.database.query.where
@@ -34,6 +40,7 @@ internal class DefaultReadService @AssistedInject constructor(@Assisted private 
                                                               private val monarchy: Monarchy,
                                                               private val taskExecutor: TaskExecutor,
                                                               private val setReadMarkersTask: SetReadMarkersTask,
+                                                              private val readReceiptsSummaryMapper: ReadReceiptsSummaryMapper,
                                                               private val credentials: Credentials
 ) : ReadService {
 
@@ -86,4 +93,16 @@ internal class DefaultReadService @AssistedInject constructor(@Assisted private 
         return isEventRead
     }
 
+    override fun getEventReadReceiptsLive(eventId: String): LiveData<List<ReadReceipt>> {
+        val liveEntity = RealmLiveData(monarchy.realmConfiguration) { realm ->
+            ReadReceiptsSummaryEntity.where(realm, eventId)
+        }
+        return Transformations.map(liveEntity) { realmResults ->
+            realmResults.firstOrNull()?.let {
+                readReceiptsSummaryMapper.map(it)
+            }?.sortedByDescending {
+                it.originServerTs
+            } ?: emptyList()
+        }
+    }
 }