+++ matrix-sdk-android/src/main/java/im/vector/matrix/android/internal/session/room/timeline/DefaultTimelineService.kt
@@ -25,8 +25,10 @@ import im.vector.matrix.android.api.session.crypto.CryptoService
 import im.vector.matrix.android.api.session.room.timeline.Timeline
 import im.vector.matrix.android.api.session.room.timeline.TimelineEvent
 import im.vector.matrix.android.api.session.room.timeline.TimelineService
+import im.vector.matrix.android.api.session.room.timeline.TimelineSettings
 import im.vector.matrix.android.internal.database.RealmLiveData
-import im.vector.matrix.android.internal.database.mapper.asDomain
+import im.vector.matrix.android.internal.database.mapper.ReadReceiptsSummaryMapper
+import im.vector.matrix.android.internal.database.mapper.TimelineEventMapper
 import im.vector.matrix.android.internal.database.model.TimelineEventEntity
 import im.vector.matrix.android.internal.database.query.where
 import im.vector.matrix.android.internal.task.TaskExecutor
@@ -37,7 +39,9 @@ internal class DefaultTimelineService @AssistedInject constructor(@Assisted priv
                                                                   private val taskExecutor: TaskExecutor,
                                                                   private val contextOfEventTask: GetContextOfEventTask,
                                                                   private val cryptoService: CryptoService,
-                                                                  private val paginationTask: PaginationTask
+                                                                  private val paginationTask: PaginationTask,
+                                                                  private val timelineEventMapper: TimelineEventMapper,
+                                                                  private val readReceiptsSummaryMapper: ReadReceiptsSummaryMapper
 ) : TimelineService {
 
     @AssistedInject.Factory
@@ -45,7 +49,7 @@ internal class DefaultTimelineService @AssistedInject constructor(@Assisted priv
         fun create(roomId: String): TimelineService
     }
 
-    override fun createTimeline(eventId: String?, allowedTypes: List<String>?): Timeline {
+    override fun createTimeline(eventId: String?, settings: TimelineSettings): Timeline {
         return DefaultTimeline(roomId,
                                eventId,
                                monarchy.realmConfiguration,
@@ -53,7 +57,10 @@ internal class DefaultTimelineService @AssistedInject constructor(@Assisted priv
                                contextOfEventTask,
                                paginationTask,
                                cryptoService,
-                               allowedTypes)
+                               timelineEventMapper,
+                               settings,
+                               TimelineHiddenReadReceipts(readReceiptsSummaryMapper, roomId, settings)
+        )
     }
 
     override fun getTimeLineEvent(eventId: String): TimelineEvent? {
@@ -61,7 +68,7 @@ internal class DefaultTimelineService @AssistedInject constructor(@Assisted priv
                 .fetchCopyMap({
                                   TimelineEventEntity.where(it, eventId = eventId).findFirst()
                               }, { entity, realm ->
-                                  entity.asDomain()
+                                  timelineEventMapper.map(entity)
                               })
     }
 
@@ -69,8 +76,8 @@ internal class DefaultTimelineService @AssistedInject constructor(@Assisted priv
         val liveData = RealmLiveData(monarchy.realmConfiguration) {
             TimelineEventEntity.where(it, eventId = eventId)
         }
-        return Transformations.map(liveData) {
-            it.firstOrNull()?.asDomain()
+        return Transformations.map(liveData) { events ->
+            events.firstOrNull()?.let { timelineEventMapper.map(it) }
         }
     }